% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chainhg38_hg19}
\alias{chainhg38_hg19}
\title{hg38tohg19 chain file}
\format{
An object of class \code{Chain} of length 25.
}
\usage{
chainhg38_hg19
}
\description{
Chain file to lift positions from hg38 to hg19. It can be used with rtracklayer::liftover function
}
\keyword{datasets}
