% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{collect_beta_panel}
\alias{collect_beta_panel}
\title{Collect Beta Panel}
\usage{
collect_beta_panel(
  cov_file,
  panel_design,
  sets_analyze = c("mCRPC_PBMC", "NE_Adeno", "rockermeth_NEvsAdeno",
    "DMS_genesOfInterest", "Promoter_DEgenes", "CpG_site_DNAmeth_classifierJCI_2020")
)
}
\arguments{
\item{cov_file}{Uniformed output file}

\item{panel_design}{Data frame with CONCATENATED design and region_set and reg_id columns}

\item{sets_analyze}{The panel submodules that should be processed}
}
\value{
A data.frame with collapsed beta values by region
}
\description{
Collect Beta Panel
}
