% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{compute_all}
\alias{compute_all}
\title{Compute PE score wrapper}
\usage{
compute_all(
  dd,
  atlas_tc,
  ref_mat,
  atlas_samples,
  cfdna_ids,
  mode = "brms",
  sequential = FALSE,
  nclust = 4,
  limit = 0.03
)
}
\arguments{
\item{dd}{A matrix of DNA methylation with informative regions on rows and samples on columns}

\item{atlas_tc}{A data.frame with previously computed tumor content estimation for each sample}

\item{ref_mat}{A matrix with mean and sd DNA methylation for each cell-type of interest and each region}

\item{atlas_samples}{List of samples for which the TC should be considered 1 regardless of the estimation}

\item{cfdna_ids}{List of samples for which healthy cfDNA should be used as background, instead of PBMCs (default)}

\item{mode}{Deconvolution strategy used, default to brms Bayesian regression (Stan backend)}

\item{sequential}{execute the computation sequentially (for debugging purpose)}

\item{nclust}{Number of cluster for parallel computation (default:4)}

\item{limit}{The lower limit of PE score estimation (Default:3\%)}
}
\value{
A data.frame with PE score
}
\description{
Compute PE score wrapper
}
