% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{compute_evidence_brms}
\alias{compute_evidence_brms}
\title{Decompose tumor component into NE and Adeno modeling observed signal as a linear combination of reference profiles,
using a strong prior based on informed tumor content estimation}
\usage{
compute_evidence_brms(
  obs,
  cf_mu,
  cf_sd,
  ad_mu,
  ad_sd,
  ne_mu,
  ne_sd,
  tc,
  tc_lw,
  tc_up
)
}
\arguments{
\item{obs}{Observed DNA methylation (Beta) for each region, Vector}

\item{cf_mu}{Mean DNA methylation for each region in cfDNA}

\item{cf_sd}{SD DNA methylation for each region in cfDNA}

\item{ad_mu}{Mean DNA methylation for each region in CRPC-Adeno}

\item{ad_sd}{SD DNA methylation for each region in CRPC-Adeno}

\item{ne_mu}{Mean DNA methylation for each region in CRPC-NE}

\item{ne_sd}{SD DNA methylation for each region in CRPC-NE}

\item{tc}{Estimation of tumor content for the current sample}

\item{tc_lw}{lower condifence of TC (not currently used)}

\item{tc_up}{upper confidence of TC (not currently used)}
}
\value{
The estimated components for the observed sample
}
\description{
Decompose tumor component into NE and Adeno modeling observed signal as a linear combination of reference profiles,
using a strong prior based on informed tumor content estimation
}
