% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{compute_ic_meth}
\alias{compute_ic_meth}
\title{Compute the fraction of non-WBC derived cell-free DNA in a sample}
\usage{
compute_ic_meth(tumor_mat, control_mat, reg_df)
}
\arguments{
\item{tumor_mat}{Matrix of samples in which estimate epithelial (tumoral) fraction, regions on rows, samples on columns.}

\item{control_mat}{Matrix of PBMC or cfDNA DNA methylation profiles, regions on rows, samples on columns.}

\item{reg_df}{Dataframe of informative regions with 2 columns reg_id and meth state (-1/+1) for hypo and hyper respectively, reporting the differential between CRPC samples and PBMC/cfDNA}
}
\value{
A vector with TC estimation for input samples
}
\description{
Compute the fraction of non-WBC derived cell-free DNA in a sample
}
