% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contigs_hg38}
\alias{contigs_hg38}
\title{contigs_hg38 chain file}
\format{
An object of class \code{data.frame} with 25 rows and 5 columns.
}
\usage{
contigs_hg38
}
\description{
Chromosome names in hg38 contig format
}
\keyword{datasets}
