% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_mat}
\alias{control_mat}
\title{control_mat_cfDNA matrix of cfDNA samples}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1886 rows and 23 columns.
}
\usage{
control_mat
}
\description{
Reference samples for tumor content estimation (from Fox-Fisher et al. eLife2021)
}
\keyword{datasets}
