% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_mat_pbmc}
\alias{control_mat_pbmc}
\title{control_mat_PBMC matrix of PBMC samples}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1886 rows and 23 columns.
}
\usage{
control_mat_pbmc
}
\description{
Reference samples for tumor content estimation in tissue (from the BLUEPRINT consortium)
}
\keyword{datasets}
