% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.clamp_val}
\alias{.clamp_val}
\title{Clamp values}
\usage{
.clamp_val(vec, lower, upper)
}
\arguments{
\item{vec}{a vector to clamp}

\item{lower}{upper bound}

\item{upper}{lower bound}
}
\value{
a clamped numeric vector
}
\description{
Clamp values
}
