% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.probit_trans}
\alias{.probit_trans}
\title{Probit transformation}
\usage{
.probit_trans(x, multiplier = 1)
}
\arguments{
\item{x}{beta values}

\item{multiplier}{the number that multiplies -x at the exponent (default is 1)}
}
\value{
probit transformed value
}
\description{
Probit transformation
}
