% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{format_score}
\alias{format_score}
\title{Compute final scores from deconvolved fractions}
\usage{
format_score(score_df, limit = 0.05)
}
\arguments{
\item{score_df}{input data.frame}

\item{limit}{lower limit of detection for tumor content}
}
\value{
A data.frame with complete estimations
}
\description{
Compute final scores from deconvolved fractions
}
