% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_design}
\alias{panel_design}
\title{panel_design chain file}
\format{
An object of class \code{data.frame} with 1981 rows and 27 columns.
}
\usage{
panel_design
}
\description{
A data.frame with the NEMO panel design, annotating regions subsets, methylation states, genes, and positions
}
\keyword{datasets}
