% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refDist}
\alias{refDist}
\title{refDist chain file}
\format{
An object of class \code{data.frame} with 1287 rows and 10 columns.
}
\usage{
refDist
}
\description{
Reference distributions for cell types of interest: ad:Adeno, ne:NE, adm:PBMC, cf:cfDNA
}
\keyword{datasets}
