% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{uniform_input}
\alias{uniform_input}
\title{Uniform input}
\usage{
uniform_input(file_in, format, all_cpg, min_cov = 10, ch_file)
}
\arguments{
\item{file_in}{path to the input file with DNA methylation counts}

\item{format}{the format in which files are expected}

\item{all_cpg}{the position of all CpGs in the human genome}

\item{min_cov}{the minimum coverage required to retain a CpG site}

\item{ch_file}{chain file for hg38 to hg19 conversion - used for datasets with hg38 data}
}
\value{
A uniformed version of the count from the target file
}
\description{
Uniform input
}
