% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_beta.R
\name{collapse_stats}
\alias{collapse_stats}
\title{Collapse statistics of regions}
\usage{
collapse_stats(res_collected, ssid)
}
\arguments{
\item{res_collected}{}
}
\value{
A data.frame of collapsed regions
}
\description{
Collapse statistics of regions
}
